Move Parsing object into separate source file
